# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quire Server API",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "description": "Private reading-data sync and EPUB storage. Account and watched-folder administration are owner CLI operations."
  },
  "paths": {
...
    "/v1/files": {
      "get": {
        "summary": "List owned file availability",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Available EPUBs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "files": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ServerFile"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error; only owned files are accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/books/{id}/file": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          }
        }
      ],
      "put": {
        "summary": "Upload an EPUB after verifying its SHA-256 identity",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/epub+zip": {
              "schema": {
                "type": "string",
                "format": "binary",
                "maxLength": 134217728
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upload registered"
          },
          "default": {
            "description": "Error; only owned files are accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Download an owned EPUB",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "EPUB",
            "content": {
              "application/epub+zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "default": {
            "description": "Error; only owned files are accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove uploaded source only; never remove watched sources or reading data",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Upload removed"
          },
          "403": {
            "description": "Only a watched source exists"
          },
          "default": {
            "description": "Error; only owned files are accessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/books/{id}/metadata": {
      "get": {
        "summary": "Get EPUB metadata and a small raster cover without downloading the book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Extracted title, author, series, nullable volume, and cover (JPEG data URL, or empty string). Cover is at most 320 by 480 pixels. EPUB file ownership is required.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "author": {
                      "type": "string"
                    },
                    "series": {
                      "type": "string"
                    },
                    "volume": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "cover": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "404": {
            "description": "Book unavailable for this account"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "ServerFile": {
        "type": "object",
        "required": [
          "bookId",
          "size",
          "uploaded",
          "watched"
        ],
        "properties": {
          "bookId": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "size": {
            "type": "integer",
            "minimum": 0,
            "maximum": 134217728
          },
          "uploaded": {
            "type": "boolean"
          },
          "watched": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}